// src/app/components/articles/types.ts
import { IArticle, ISortDirection } from '@src/generated/types';
import { GraphQLError } from 'graphql';
import { Observable, ObservableInput } from 'rxjs';

export interface IObservables extends Record<string, ObservableInput<unknown>> {
    loading: Observable<boolean>;
    errors: Observable<readonly GraphQLError[] | undefined>;
    articles: Observable<IArticle[]>;
    sortUpdatedAt: Observable<ISortDirection>;
}